starting front end setup

+++ run.sh
@@ -1,10 +1,15 @@
+if [ $# -ne 2]; then
+  echo "$0 env(dev|prod) action (apply|destroy)"
+  exit 1
+fi
 git pull
 rm -rf .terraform
-terraform init -backend-config=env-$1/state.tfvars
-terraform apply -var-file=env-$1/main.tfvars -auto-approve
+terraform init -backend-config=env-${1}/state.tfvars
+terraform $2 -var-file=env-${1}/main.tfvars -auto-approve
 
 
 
 
 
 #$1 = dev, test, prod etc...
+#$2 = plan, apply, destroy etc...
